pyproject.toml
[project]
name = "sky-habbits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0",
    "django-celery-beat>=2.7.0",
    "django-cors-headers>=4.4.0",
    "django-types>=0.19.1",
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "drf-social-oauth2>=3.1.0",
    "drf-yasg>=1.21.7",
    "psycopg[binary]>=3.2.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "setuptools>=75.1.0",
]
